Refactoring Cards
Moved utility functions to a utils file to help with file length and modularization
